Added Beta Art 1
Added all art assests I current to Untiy, Started to add some to
bedroom. Stopped because I am getting a lot of errors.

"NullReferenceException: Object reference not set to an instance of an
object
ItemGetNeedEvent.Update () (at
Assets/Scripts/ItemGetNeedEvent.cs:16)"
On the collectibles and the laptop, don't think the art broke it, got
the error when I opened the scene in Unity.